move utilties to SpannerUtils.java, delete unused code
