--- spark-3.1-spanner-lib/src/main/java/com/google/cloud/spark/spanner/SpannerPartitionReader.java
@@ -1,60 +1,0 @@
-// Copyright 2023 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package com.google.cloud.spark.spanner;
-
-import com.google.cloud.spanner.BatchClient;
-import com.google.cloud.spanner.BatchReadOnlyTransaction;
-import com.google.cloud.spanner.Partition;
-import com.google.cloud.spanner.ResultSet;
-import com.google.cloud.spanner.TimestampBound;
-import java.io.IOException;
-import org.apache.spark.sql.Row;
-import org.apache.spark.sql.connector.read.InputPartition;
-import org.apache.spark.sql.connector.read.PartitionReader;
-
-public class SpannerPartitionReader<Row> implements PartitionReader<Row> {
-
-  private final ResultSet rs;
-  private BatchReadOnlyTransaction txn;
-
-  public <Row> SpannerPartitionReader(BatchClient batchClient, InputPartition sparkPartition) {
-    // TODO: Confirm if holding the transaction for longer
-    // than 10 seconds causes it to time out and if perhaps
-    // we can open up a longer standing transaction.
-    this.txn = batchClient.batchReadOnlyTransaction(TimestampBound.strong());
-    Partition spannerPartition = (Partition) sparkPartition;
-    this.rs = txn.execute(spannerPartition);
-  }
-
-  @Override
-  public boolean next() {
-    return this.rs.next();
-  }
-
-  @Override
-  public Row get() {
-    return SpannerSpark.resultSetRowToSparkRow(this.rs);
-  }
-
-  @Override
-  public void close() throws IOException {
-    this.rs.close();
-    this.txn.close();
-  }
-
-  public boolean supportsColumnarReads(InputPartition partition) {
-    return false;
-  }
-}
