--- spark-3.1-spanner-lib/src/main/java/com/google/cloud/spark/spanner/SpannerPartitionReaderFactory.java
@@ -1,45 +1,0 @@
-// Copyright 2023 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package com.google.cloud.spark.spanner;
-
-import com.google.cloud.spanner.BatchClient;
-import java.util.Map;
-import org.apache.spark.sql.catalyst.InternalRow;
-import org.apache.spark.sql.connector.read.InputPartition;
-import org.apache.spark.sql.connector.read.PartitionReader;
-import org.apache.spark.sql.connector.read.PartitionReaderFactory;
-import org.apache.spark.sql.vectorized.ColumnarBatch;
-
-/*
- * SpannerPartitionReaderFactory is an entry to implement PartitionReaderFactory.
- */
-public class SpannerPartitionReaderFactory implements PartitionReaderFactory {
-  private BatchClient batchClient;
-
-  public SpannerPartitionReaderFactory(Map<String, String> opts) {
-    this.batchClient = SpannerUtils.batchClientFromProperties(opts);
-  }
-
-  @Override
-  public PartitionReader<ColumnarBatch> createColumnarReader(InputPartition partition) {
-    // TODO: Fill me in.
-    return null;
-  }
-
-  @Override
-  public PartitionReader<InternalRow> createReader(InputPartition partition) {
-    return new SpannerPartitionReader(batchClient, partition);
-  }
-}
